更新小问题 修改tcp的连接响应方式。 file chunk struct changed store interface update
finish slice debug one step...
SubscribeOn(Schedules.boundedElastic)
update name
create test module
1.debug single file self upload finish
2.debug single file self download finish range,if-range
continue:
1.single slice upload all finish,
2.dft command truansfer
3.master elect
4.upload
5.download

+++ common/src/main/java/com/ly/common/domain/FileChunkState.java
@@ -71,7 +71,7 @@ public class FileChunkState {
             return -1;
         }
         if (((states[i] >>> o) & 1) > 0) {
-            states[i] &= ~(1 << o);
+            states[i] &= (byte) ~(1 << o);
             return count.addAndGet(-1);
         } else {
             return getCount();
@@ -87,7 +87,7 @@ public class FileChunkState {
         } else {
             int i = index / 8 + 4;
             int o = index % 8;
-            states[i] |= 1 << o;
+            states[i] |= (byte) (1 << o);
             return incrementCountAndGet();
         }
     }
